move show karma personally into handlers
